+++ build_modules/lib/src/module_library_builder.dart
@@ -8,21 +8,23 @@ import 'package:build/build.dart';
 
 import 'module_library.dart';
 
-/// Creates `.library` assets listing the dependencies and parts for aDart
-/// library, as well as whether it is an entrypoint.
+const moduleLibraryExtension = '.module.library';
+
+/// Creates `.module.library` assets listing the dependencies and parts for a
+/// Dart library, as well as whether it is an entrypoint.
 ///
 ///
 /// The output format is determined by [ModuleLibrary.toString] and can be
 /// restored by [ModuleLibrary.parse].
 ///
-/// Non-importable Dart source files will not get a `.library` asset output. See
-/// [ModuleLibrary.isImportable].
+/// Non-importable Dart source files will not get a `.module.library` asset
+/// output. See [ModuleLibrary.isImportable].
 class ModuleLibraryBuilder implements Builder {
   const ModuleLibraryBuilder();
 
   @override
   final buildExtensions = const {
-    '.dart': const ['.dart.library']
+    '.dart': const [moduleLibraryExtension]
   };
 
   @override
@@ -31,6 +33,6 @@ class ModuleLibraryBuilder implements Builder {
         buildStep.inputId, await buildStep.readAsString(buildStep.inputId));
     if (!library.isImportable) return;
     await buildStep.writeAsString(
-        buildStep.inputId.addExtension('.library'), '$library');
+        buildStep.inputId.changeExtension(moduleLibraryExtension), '$library');
   }
 }